+++ geostore/tests/test_models/test_other_models.py
@@ -1,3 +1,4 @@
+from django.db.utils import IntegrityError
 from django.test import TestCase
 from django.utils.text import slugify
 
@@ -34,3 +35,13 @@ class FeatureTestCase(TestCase):
                                          })
         feature.clean()
         self.assertIsNotNone(feature.pk)
+
+    def test_constraint_feature_empty_geom(self):
+        with self.assertRaises(IntegrityError):
+            Feature.objects.create(layer=self.layer_schema,
+                                   geom='POINT EMPTY',)
+
+    def test_constraint_feature_valid_geom(self):
+        with self.assertRaises(IntegrityError):
+            Feature.objects.create(layer=self.layer_schema,
+                                   geom='POLYGON((0 0, 1 1, 1 2, 1 1, 0 0))',)