+++ geostore/tests/test_models/test_other_models.py
@@ -1,9 +1,11 @@
+from django.contrib.gis.geos import Point
 from django.db.utils import IntegrityError
 from django.test import TestCase
 from django.utils.text import slugify
 
 from geostore import GeometryTypes
 from geostore.models import LayerExtraGeom, Feature
+from geostore import settings as app_settings
 from geostore.tests.factories import LayerSchemaFactory
 
 
@@ -36,6 +38,14 @@ class FeatureTestCase(TestCase):
         feature.clean()
         self.assertIsNotNone(feature.pk)
 
+    def test_feature_geom_3d_to_2d(self):
+        feature = Feature.objects.create(layer=self.layer_schema,
+                                         geom='POINT(0 0 0)',
+                                         properties={
+                                             'name': 'toto'
+                                         })
+        self.assertEqual(feature.geom, Point(0, 0, srid=app_settings.INTERNAL_GEOMETRY_SRID))
+
     def test_constraint_feature_empty_geom(self):
         with self.assertRaises(IntegrityError):
             Feature.objects.create(layer=self.layer_schema,
@@ -45,3 +55,4 @@ class FeatureTestCase(TestCase):
         with self.assertRaises(IntegrityError):
             Feature.objects.create(layer=self.layer_schema,
                                    geom='POLYGON((0 0, 1 1, 1 2, 1 1, 0 0))',)
+