+++ CHANGES.md
@@ -4,7 +4,8 @@ CHANGELOG
 0.4.4.dev0     (XXXX-XX-XX)
 ---------------------------
 
-* add relation property to handle relations directly in feature
+* Add relation property to handle relations directly in feature
+* Force geometries 2d
 
 
 0.4.3          (2020-10-09)