+++ CHANGES.md
@@ -4,9 +4,9 @@ CHANGELOG
 0.4.4.dev0     (XXXX-XX-XX)
 ---------------------------
 
-* Add relation property to handle relations directly in feature
 * Add constraint empty geometries
 * Add validation of constraints on geometries
+* Improve tile generation
 
 
 0.4.3          (2020-10-09)