+++ CHANGES.md
@@ -4,7 +4,8 @@ CHANGELOG
 0.4.4.dev0     (XXXX-XX-XX)
 ---------------------------
 
-* add relation property to handle relations directly in feature
+* Add constraint empty geometries
+* Add validation of constraints on geometries
 * Improve tile generation