+++ CHANGES.md
@@ -4,6 +4,8 @@ CHANGELOG
 0.4.4.dev0     (XXXX-XX-XX)
 ---------------------------
 
+* Add relation property to handle relations directly in feature
+* Force geometries 2d
 * Add constraint empty geometries
 * Add validation of constraints on geometries
 * Improve tile generation